requirements.txt
colorama
rich
prompt_toolkit
bcrypt
argon2-cffi
requests
dnspython
python-whois
PyYAML
